+++ main.py
@@ -2,7 +2,7 @@ import numpy as np
 from cmath import phase
 from component import LoadCurrent
 from power_network import PowerNetwork
-from bus import BusSlack, BusPV, BusPQ, Bus
+from bus.bus import BusSlack, BusPV, BusPQ, Bus
 from branch import BranchPi
