+++ power_network.py
@@ -2,7 +2,7 @@ import numpy as np
 from scipy.optimize import root
 from  cmath import phase
 
-from bus import Bus
+from bus.bus import Bus
 from branch import Branch
 
 class PowerNetwork():
@@ -94,11 +94,10 @@ class PowerNetwork():
 
     def set_equilibrium(self, V, I):
         for idx in range(len(self.a_bus)):
-            self.a_bus[idx].set_equilibrium(V[idx],I[idx])
+            self.a_bus[idx].set_equilibrium(V[idx][0],I[idx][0])
 
     def initialize(self):
         V, I = self.calculate_power_flow()
-
         self.set_equilibrium(V, I)
