+++ zz_check_generator_out.py
@@ -1,7 +1,7 @@
 import pandas as pd
 from numpy import pi
 
-from generator_1axis import Generator1Axis
+from generator.generator_1axis import Generator1Axis
 
 omega0 = 60*2*pi
 mac = {'Xd':1.569, 'Xd_prime':0.963, 'Xq':0.963, 'T':5.14, 'M':100, 'D':10}
